feat: Add Media, Notification, Respect Modes, and Speech Template settings sections
- Implemented MediaBehaviorSection to manage media behavior settings including pause, duck, silence, and ignore options.
- Created NotificationBehaviorSection for handling notification behaviors with priority app management and different modes (interrupt, queue, skip, smart).
- Added RespectModesSection to respect device modes like Do Not Disturb, Silent, and Vibrate, along with notification deduplication and dismissal memory settings.
- Added SpeechTemplateSection for customizing speech templates with various presets and a custom input option.
- Refactored SettingsDatabase to include references to new BehaviorSettingsActivity.
