+++ app/src/github/java/com/micoyc/speakthat/SettingsDatabase.kt
@@ -1,6 +1,7 @@
 package com.micoyc.speakthat
 
 import android.content.Intent
+import com.micoyc.speakthat.settings.BehaviorSettingsActivity
 
 object SettingsDatabase {
