+++ app/src/main/java/com/micoyc/speakthat/PriorityAppAdapter.kt
@@ -7,6 +7,7 @@ import android.view.ViewGroup
 import android.widget.ImageButton
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
+import com.micoyc.speakthat.settings.BehaviorSettingsActivity
 
 class PriorityAppAdapter(
     private val items: List<String>,
