+++ app/src/main/java/com/micoyc/speakthat/settings/BehaviorSettingsSection.java
@@ -1,0 +1,7 @@
+package com.micoyc.speakthat.settings;
+
+public interface BehaviorSettingsSection {
+    void bind();
+    void load();
+    void release();
+}
